Add postdebug to post gist/issue with diagnostics.
Add postdebug command to ~/bin/swgemu.

It can ask user if they would like to open an issue or if they want to
post a gist.  They will be prompted to login if they want an issue and
optionally they can post a gist w/o logging in to post an annonymous
gist for sharing.

Provides the URL after it runs so the user can share in IRC or via some
other mechanism to get help.